Store offset and point count in octree datastructure during subsampling

+++ OctreeApi/OctreeLibrary/src/include/types.cuh
@@ -6,14 +6,11 @@
 
 struct SubsampleConfig
 {
-    // uint32_t* lutAdress;
-    // uint32_t lutStartIndex;
-    // uint64_t* averagingAdress;
     uint32_t linearIdx;
     int sparseIdx;
     bool isParent;
     uint32_t leafPointAmount; // Only valid if isParent == false
-    uint32_t leafDataIdx;     // Only valid if isParent == false
+    uint64_t leafDataIdx;     // Only valid if isParent == false
 };
 
 struct __align__ (16) SubsampleSet
@@ -33,7 +30,7 @@ struct Chunk
     uint32_t pointCount;       // How many points does this chunk have
     uint32_t parentChunkIndex; // Determines the INDEX of the parent CHUNK in the GRID - Only needed during Merging
     bool isFinished;           // Is this chunk finished (= not mergeable anymore)
-    uint32_t chunkDataIndex;   // Determines the INDEX in the chunk data array -> for storing point data
+    uint64_t chunkDataIndex;   // Determines the INDEX in the chunk data array -> for storing point data
     int childrenChunks[8];     // The INDICES of the children chunks in the GRID
     bool isParent;             // Denotes if Chunk is a parent or a leaf node
 };
